scripts/requirements.txt
requests
python-dotenv
prometheus_client
flask
discord